requirements.txt
numpy
pandas
matplotlib
tensorflow
pymongo
scikit-learn
python_deriv_api
websocket-client
mongoengine
dash
plotly
